Move the set_size test to the shared backend tests
The test-backends crate runs each case under both the winit and the Qt
backend, in a separate process, with the backend set explicitly.
