+++ internal/backends/winit/Cargo.toml
@@ -161,7 +161,3 @@ rustdoc-args = ["--generate-link-to-definition"]
 [[test]]
 name = "menubar_borrow"
 harness = false
-
-[[test]]
-name = "set_size_before_show"
-harness = false
